refactor(qnn): remove unused output order tracking and add debug utilities
- Remove qnnOutputNameToIndex_ map and related functionality that was
  tracking QNN output order since it's no longer needed
- Remove expectedOutputOrder_ and associated getter/setter methods
  from QNNModel class
- Remove the logic in QNNGraphBuildPass that was recording MLLM
  expected output order from ReturnOp
- Add new debug utility functions for printing dequantized UInt16
  tensor data with support for multi-dimensional arrays
- Fix integer casting issue in PromptProcessor prefill method

+++ mllm/backends/qnn/aot_rt/PromptProcessor.cpp
@@ -148,7 +148,7 @@ int64_t PromptProcessor<T>::prefill(const std::vector<int64_t>& prompt_tokens, i
     current_pos += chunk_size;
   }
 
-  auto logits = output_tensors_[0].to(kCPU).squeeze(0)[{kAll, (num_tokens + config_.ar_len - 1) % config_.ar_len, kAll}];
+  auto logits = output_tensors_[0].to(kCPU).squeeze(0)[{kAll, ((int)num_tokens + config_.ar_len - 1) % config_.ar_len, kAll}];
 
   auto cur_token = module_->sampleGreedy(logits);
